+++ python/examples/dualarmdemo_schunk.py
@@ -181,8 +181,7 @@ def run(args=None):
                       action="store",type='string',dest='scene',default='data/dualarmmanipulation.env.xml',
                       help='Scene file to load')   
     (options, leftargs) = parser.parse_args(args=args)
-    env = OpenRAVEGlobalArguments.parseAndCreate(options,defaultviewer=True)    
-    main(env,options)
+    OpenRAVEGlobalArguments.parseAndCreateThreadedUser(options,main,defaultviewer=True)
 
 if __name__ == "__main__":
     run()