+++ python/examples/fastgraspingthreaded.py
@@ -215,8 +215,7 @@ def run(args=None):
     parser.add_option('--manipname', action="store",type='string',dest='manipname',default=None,
                       help='Choose the manipulator to perform the grasping for')
     (options, leftargs) = parser.parse_args(args=args)
-    env = OpenRAVEGlobalArguments.parseAndCreate(options,defaultviewer=True)
-    main(env,options)
+    OpenRAVEGlobalArguments.parseAndCreateThreadedUser(options,main,defaultviewer=True)
 
 if __name__ == "__main__":
     run()