+++ python/examples/movehandstraight.py
@@ -121,8 +121,7 @@ def run(args=None):
                       action="store",type='string',dest='manipname',default=None,
                       help='Choose the manipulator to perform movement for')
     (options, leftargs) = parser.parse_args(args=args)
-    env = OpenRAVEGlobalArguments.parseAndCreate(options,defaultviewer=True)
-    main(env,options)
+    OpenRAVEGlobalArguments.parseAndCreateThreadedUser(options,main,defaultviewer=True)
 
 if __name__ == "__main__":
     run()