+++ python/examples/testphysics_controller.py
@@ -63,8 +63,7 @@ def run(args=None):
     parser.add_option('--scene',action="store",type='string',dest='scene',default='data/hanoi.env.xml',
                       help='Scene file to load (default=%default)')
     (options, leftargs) = parser.parse_args(args=args)
-    env = OpenRAVEGlobalArguments.parseAndCreate(options,defaultviewer=True)
-    main(env,options)
+    OpenRAVEGlobalArguments.parseAndCreateThreadedUser(options,main,defaultviewer=True)
 
 if __name__=='__main__':
     run()