+++ python/examples/testviewercallback.py
@@ -61,8 +61,7 @@ def run(args=None):
                       action="store",type='string',dest='scene',default='data/lab1.env.xml',
                       help='OpenRAVE scene to load')
     (options, leftargs) = parser.parse_args(args=args)
-    env = OpenRAVEGlobalArguments.parseAndCreate(options,defaultviewer=True)
-    main(env,options)
+    OpenRAVEGlobalArguments.parseAndCreateThreadedUser(options,main,defaultviewer=True)
 
 if __name__=='__main__':
     run()