+++ python/examples/tutorial_grasptransform.py
@@ -203,8 +203,7 @@ def run(args=None):
     parser = OptionParser(description='Find the transform that moves the hand to target')
     OpenRAVEGlobalArguments.addOptions(parser)
     (options, leftargs) = parser.parse_args(args=args) # use default options 
-    env = OpenRAVEGlobalArguments.parseAndCreate(options,defaultviewer=True) # the special setup for openrave tutorial
-    main(env,options)
+    OpenRAVEGlobalArguments.parseAndCreateThreadedUser(options,main,defaultviewer=True)
 
 if __name__ == '__main__':
     run()