+++ python/examples/visibilityplanning.py
@@ -426,8 +426,7 @@ def run(args=None):
     parser.add_option('--nocameraview',action="store_false",dest='usecameraview',default=True,
                       help='If set, will not open any camera views')
     (options, leftargs) = parser.parse_args(args=args)
-    env = OpenRAVEGlobalArguments.parseAndCreate(options,defaultviewer=True)
-    main(env,options)
+    OpenRAVEGlobalArguments.parseAndCreateThreadedUser(options,main,defaultviewer=True)
 
 if __name__=='__main__':
     run()